LayoutTests/imported/w3c/web-platform-tests/css/css-tables/width-distribution/computing-column-measure-1-expected.txt

Height Distribution

"Computing row measures"

This is testing intermediate min-content height for span 2


PASS Checking intermediate min-content height for span 2 (1)
PASS Checking intermediate min-content height for span 2 (2)
PASS Checking intermediate min-content height for span 2 (3)
FAIL Checking intermediate min-content height for span 2 (4) assert_equals: expected 17 but got 10
